+++ src/mokeup/favorList.json
@@ -1,0 +1,26 @@
+[
+  {
+    "id": 1,
+    "name": "LOCA 365 카드",
+    "company": "롯데카드",
+    "image": "https://api.card-gorilla.com:8080/storage/card/2330/card_img/24131/2330card.png"
+  },
+  {
+    "id": 2,
+    "name": "삼성카드 taptap O",
+    "company": "삼성카드",
+    "image": "https://api.card-gorilla.com:8080/storage/card/51/card_img/27707/51card.png"
+  },
+  {
+    "id": 3,
+    "name": "DA@카드의정석",
+    "company": "우리카드",
+    "image": "https://api.card-gorilla.com:8080/storage/card/87/card_img/20239/87card.png"
+  },
+  {
+    "id": 4,
+    "name": "신한카드 Mr.Life",
+    "company": "신한카드",
+    "image": "https://api.card-gorilla.com:8080/storage/card/13/card_img/20095/13card.png"
+  }
+]
